requirements.txt
click==8.1.3
dj_static==0.0.6
django_celery_beat==2.4.0
emails==0.6
ems==0.2.25
future==0.18.2
fuzzywuzzy==0.18.0
ldap3==2.9.1
pandas==1.5.1
pyTelegramBotAPI==4.7.1
python_dateutil==2.8.2
pytz==2022.6
redis==4.3.4
requests==2.28.1
waitress==2.1.2
whitenoise==6.2.0
xlrd==2.0.1
